Refactor document management operations
- Renamed the `documents` module to `document_service` for clarity and consistency.
- Updated import statements across the codebase to reflect the new module name.
- Introduced a new `document_service.py` file containing the `DocumentService` class, which handles PostgreSQL CRUD operations with OpenSearch synchronization.
- Removed the old `__init__.py` file from the `documents` module.
- Added unit tests for the new `DocumentService` class to ensure functionality and reliability.
