--- robosystems/operations/documents/__init__.py
@@ -1,5 +1,0 @@
-"""Document management operations."""
-
-from .service import DocumentService
-
-__all__ = ["DocumentService"]
